% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files.R.R
\name{write_table_rj}
\alias{write_table_rj}
\title{Write table to clipboard}
\usage{
write_table_rj(df)
}
\arguments{
\item{df}{Data frame that you want to write to the clipboard.}
}
\value{
No return: Data frame is in the clipboard.
}
\description{
This function writes a data frame to a format that can be easily pasted into Excel.
}
